+++ Block/BlockRendererInterface.php
@@ -11,16 +11,16 @@
 
 namespace Sonata\BlockBundle\Block;
 
-use Sonata\BlockBundle\Block\BlockExecutionContextInterface;
+use Sonata\BlockBundle\Block\BlockContextInterface;
 use Symfony\Component\HttpFoundation\Response;
 
 interface BlockRendererInterface
 {
     /**
-     * @param BlockExecutionContextInterface $name
+     * @param BlockContextInterface $name
      * @param null|Response                  $response
      *
      * @return Response
      */
-    public function render(BlockExecutionContextInterface $name, Response $response = null);
+    public function render(BlockContextInterface $name, Response $response = null);
 }