+++ Block/TraceableBlockRenderer.php
@@ -41,7 +41,7 @@ class TraceableBlockRenderer implements BlockRendererInterface
     /**
      * {@inheritdoc}
      */
-    public function render(BlockExecutionContextInterface $blockContext, Response $response = null)
+    public function render(BlockContextInterface $blockContext, Response $response = null)
     {
         $this->startTracing($blockContext->getBlock());