+++ Tests/Block/TraceableBlockRendererTest.php
@@ -13,7 +13,7 @@ namespace Sonata\BlockBundle\Tests\Block;
 
 use Sonata\BlockBundle\Block\TraceableBlockRenderer;
 use Symfony\Component\HttpFoundation\Response;
-use Sonata\BlockBundle\Block\BlockExecutionContext;
+use Sonata\BlockBundle\Block\BlockContext;
 
 class TraceableBlockRendererTest extends \PHPUnit_Framework_TestCase
 {
@@ -28,7 +28,7 @@ class TraceableBlockRendererTest extends \PHPUnit_Framework_TestCase
         $block->expects($this->any())->method('getId')->will($this->returnValue(42));
         $block->expects($this->any())->method('getType')->will($this->returnValue('mytype'));
 
-        $blockContext = new BlockExecutionContext($block);
+        $blockContext = new BlockContext($block);
 
         $traceable->render($blockContext);