+++ CHANGELOG.md
@@ -9,4 +9,4 @@ A [BC BREAK] means the update will break the project for many reasons:
 
 ### 2013-03-28
 
-* [BC BREAK] Intoduction of an BlockExecutionContext to hold a BlockInterface instance and the related settings.
+* [BC BREAK] Intoduction of an BlockContext to hold a BlockInterface instance and the related settings.